# Conflicts:
#	src/main/java/com/dws/web/Event/EventRESTController.java
#	src/main/java/com/dws/web/Security/CustomRestSecurityConfig.java

+++ src/main/java/com/dws/web/Customer/CustomerRESTController.java
@@ -5,6 +5,7 @@ import com.dws.web.Event.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Collection;
@@ -28,20 +29,21 @@ public class CustomerRESTController {
     }
 
     @DeleteMapping("/admin/customer/delete/{email}")
-    public ResponseEntity<Customer> deleteCustomerAPI(@PathVariable String email) {
+    public ResponseEntity<Customer> deleteCustomerAPI(@PathVariable String email, Authentication auth) {
         Customer customer= customerService.getCustomer(email);
-        Customer c=customerService.deleteCustomer(customer.getIdClient());
-        if (c != null) {
+        Customer admin= customerService.getCustomer(auth.getName());
+        if (customer != null&&customerService.esAdmin(admin)) {
+            Customer c=customerService.deleteCustomer(customer.getIdClient());
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
-    @PutMapping("/admin/customer/update/{email}")
-    public ResponseEntity<Customer> updateCustomerAPI(@PathVariable String email, @RequestBody Customer updatedCustomer) {
+    @PutMapping("/user/customer/update/{email}")
+    public ResponseEntity<Customer> updateCustomerAPI(@PathVariable String email, @RequestBody Customer updatedCustomer, Authentication auth) {
         Customer c=customerService.getCustomer(email);
-        if (c != null) {
+        if (c != null&&auth.getName().equals(email)&&!customerService.containsCustomer(updatedCustomer)) {
             updatedCustomer.setIdClient(c.getIdClient());
             customerService.addUpdatedClient(updatedCustomer);
             return new ResponseEntity<>(updatedCustomer, HttpStatus.OK);
@@ -74,22 +76,26 @@ public class CustomerRESTController {
     //PLANNING
 
     @PostMapping("/events/{idEvent}/new/{email}")
-    public ResponseEntity<Event> newEventAPI(@PathVariable long idEvent, @PathVariable String email) {
+    public ResponseEntity<Event> newEventAPI(@PathVariable long idEvent, @PathVariable String email, Authentication auth) {
         Customer c= customerService.getCustomer(email);
         Event e=eventService.getEvent(idEvent);
-        if (c.addToPlanning(e)){
+        if (c.addToPlanning(e)&&auth.getName().equals(email)&&e!=null){
             e.assignCustomer(c);
             eventService.addEvent(e);
+            return new ResponseEntity<>(e, HttpStatus.CREATED);
         }
-        return new ResponseEntity<>(e, HttpStatus.CREATED);
+        else{
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
     }
 
     @DeleteMapping("/user/planning/{email}/delete/{idEvent}")
-    public ResponseEntity<Event> deleteEventAPI(@PathVariable String email, @PathVariable long idEvent) {
+    public ResponseEntity<Event> deleteEventAPI(@PathVariable String email, @PathVariable long idEvent, Authentication auth) {
         Customer c= customerService.getCustomer(email);
         Event e=eventService.getEvent(idEvent);
-        Event event = customerService.deleteEventFromPlanning(c, e);
-        if (event != null) {
+        if (e != null&&auth.getName().equals(email)) {
+            Event event = customerService.deleteEventFromPlanning(c, e);
             return new ResponseEntity<>(event, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -112,10 +118,10 @@ public class CustomerRESTController {
      */
 
     @GetMapping("/user/planning/{email}")
-    public ResponseEntity<Collection> getAllEventOfACustomerAPI(@PathVariable String email) {
+    public ResponseEntity<Collection> getAllEventOfACustomerAPI(@PathVariable String email, Authentication auth) {
         Customer c=customerService.getCustomer(email);
         Collection<Event> events = customerService.getAllEventsOfACustomer(c);
-        if (!events.isEmpty()) {
+        if (!events.isEmpty()&&auth.getName().equals(email)) {
             return new ResponseEntity<>(events, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -123,10 +129,10 @@ public class CustomerRESTController {
     }
 
     @GetMapping("/user/events/{email}/{idEvent}")  //Products by id
-    public ResponseEntity<Event> getEventAPI(@PathVariable String email, @PathVariable long idEvent){
+    public ResponseEntity<Event> getEventAPI(@PathVariable String email, @PathVariable long idEvent, Authentication auth){
         Customer c= customerService.getCustomer(email);
         Event e= customerService.getAnEvent(c, idEvent);
-        if (e!=null){
+        if (e!=null&&auth.getName().equals(email)){
             return new ResponseEntity<>(e, HttpStatus.OK);
         }
         else{
@@ -135,8 +141,8 @@ public class CustomerRESTController {
     }
 
     @GetMapping("/user/planning/{email}/category/{category}")
-    public ResponseEntity<Collection> planningFilteredByCategoryAPI(@PathVariable String email, @PathVariable String category) {
-        if (category.equalsIgnoreCase("ocio")||category.equalsIgnoreCase("restaurante")||category.equalsIgnoreCase("turismo")) {
+    public ResponseEntity<Collection> planningFilteredByCategoryAPI(@PathVariable String email, @PathVariable String category, Authentication auth) {
+        if ((category.equalsIgnoreCase("ocio")||category.equalsIgnoreCase("restaurante")||category.equalsIgnoreCase("turismo"))&&auth.getName().equals(email)) {
             Customer c=customerService.getCustomer(email);
             return new ResponseEntity<>(eventService.getEventsFilteredByCategory(category), HttpStatus.OK);
         }
